Vendor Goal source and embed
Creates a separate vendor-goal directory with a README explaining the
usage (and license) of vendored Goal code. Replaces the inlined strings
of Goal program source in the Go files with `//go:embed` directives.

Initially considered having Goal be a submodule or subtree, but given I
only need a small fraction of what is in that repository, I've chosen
this approach for now.