Fix raw REPL bug scanning comments
Before this fix, the raw REPL would not properly ignore comment content,
such that delimiters in comments would break scanning.

Since ari's raw REPL is a near identical copy of Goal's REPL, I filed a
bug with the Goal project and anaseto promptly fixed it. The contents of
this commit are ported directly from the commit in the Goal repo that
fixes the bug:

ec3e8a97179fd6ff8bfe035504cf0a9b506312c9

For reference, here is the Goal bug report:

https://codeberg.org/anaseto/goal/issues/18