Remove second test of time.zonename
The test two lines above shows that it does return a value for the UTC
timezone. In CI, the time.zonename value is empty for the -05:00 offset,
I assume because there is some system files/values that are missing in
the image used. On my local machine, even though it is currently EDT,
this function reported EST.