Script the release process to sync versions
Goal now supports extensions (like Ari) setting a version from Go code,
and the Go module publishing process expects a Git tag representing the
version to be used both in the Git repository and when publishing via
the 'go list' command.

This script (written in 99% Goal, but relying on a FILE binding that Ari
adds) automates the process of minting a new version, including writing
a new version.go file with a Go constant for the version, then
committing everything via Git and adding a Git tag, pushing to the
remote and then running 'go list'.