Add WASM build
Based on Goal's. The given index.html is changed slightly to show the
version (which shows that ari is loaded in the Goal evaluation context)
and adjust styling a little.

This commit also separates those Goal verbs defined in Ari that
currently work in all (universal) environments and those which do not,
to prevent trying to use verbs that will only break when called via
WebAssembly.