Auto-invoke mode with ari.c global
I have used this pattern in the past when developing an interactive CLI
REPL tool for a RESTful API. I've decided to bake it into ari as a
general tool.

ari.c stands for "context" or "current" and represents a notion of the
"current working data", an implicit contextual "here" for an interactive
program. It lacks any notion of lineage (like a "current working
directory" would have, in terms of parents or children), it's just data
in a Goal global. It can be set normally via ::

The new )goal.invoke mode mirrors what )goal provides, but additionally
interrogates the return value from evaluation. If that value is a
function-like value, it will automatically invoke it (via Apply) with
the ari.c value as argument.

In this mode, invoking a monadic function requires just writing its
name. For functions of two or more arguments, it is trivial in Goal to
create a projection where the first argument is ommitted.