Remove rich REPL
In practice, I almost never use the rich REPL. I'm almost always inside
Neovim, splitting it such that a Goal file is on the left-hand side and
a terminal is on the right-hand side. I then have been starting ari with
the --raw option enabled, and have a minimal Neovim config set up to
allow me to press Enter to send either the current line or current
selection from the left-hand Goal file to the ari REPL running on the
right.

So to make ari lighter weight, easier to develop, and focused on the
use-cases I actually have, this removes the rich REPL and all
dependencies from the code base. There were also bugs with the user
experience of that REPL (auto-complete was janky in several respects)
that I was not prepared to debug and fix.